requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
nltk>=3.6.0
textblob>=0.15.0
transformers>=4.10.0
torch>=1.9.0
shap>=0.40.0
lime>=0.2.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
google-generativeai>=0.1.0
dspy-ai>=3.0.0